Satellite Module/requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
jupyter==1.0.0
notebook==7.0.4

# Geospatial
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.1
fiona==1.9.5
pyproj==3.6.1
geemap==0.29.6
earthengine-api==0.1.368

# Satellite data
sentinelhub==3.9.1
pystac-client==0.7.3
planetary-computer==0.5.1
odc-stac==0.4.1

# Machine Learning
scikit-learn==1.3.0
scikit-image==0.21.0
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
joblib==1.3.2